feat(networking) working fine in debugger mode
