+++ foo.js
@@ -1,6 +1,5 @@
 const fetch = require('node-fetch');
-const http = require('http');
 
 fetch('http://jsonplaceholder.typicode.com/todos/5')
-  .then(response => response.json())
-  .then(json => console.log(json))
+    .then(response => response.json())
+    .then(json => console.log(json))
